Step migration storage ok

+++ lib/widgets/transaction_tile.dart
@@ -27,7 +27,7 @@ class TransactionTile extends StatelessWidget {
   final double avatarSize;
   final Transaction transaction;
   final String dateForm;
-  final int finalAmount;
+  final BigInt finalAmount;
   final DuniterIndexer duniterIndexer;
   final BuildContext context;
